requirements.txt
python-telegram-bot==12.4.2
pytest==5.4.1
redis==3.4.1
pytest-cover==3.0.0